Update DNASA.py
added all the amino acid codons to my dictionary and decided to use the RNA codons instead of DNA (simpler than differentiating b/t template or coding strands). also changed the rna translation function to translate depending on if the DNA is template or coding. started the amino acid function